Cleaning up some compiler warnings
Trying to address some possible Mac/Linux bugs

Splitting more effects into cpp/hpp
