+++ src/noise-plethora/plugins/NoisePlethoraPlugin.hpp
@@ -1,8 +1,11 @@
 #pragma once
 
 #include <rack.hpp>
+#include <memory>
+#include <string> // string might not be allowed
+#include <functional>
+#include <map>
 
-#include "PluginFactory.hpp"
 #include "../teensy/TeensyAudioReplacements.hpp"
 
 
@@ -18,7 +21,7 @@ public:
 	virtual void init() {};
 	// equivelent to arduino void loop()
 	virtual void process(float k1, float k2) {};
-	
+
 	// called once-per sample, will consume a buffer of length AUDIO_BLOCK_SAMPLES (128)
 	// then request that the buffer be refilled, returns values in range [-1, 1]
 	float processGraph() {
@@ -41,6 +44,47 @@ protected:
 	TeensyBuffer blockBuffer;
 };
 
-#define REGISTER_PLUGIN(NAME) \
-	static Registrar<NAME> NAME ##_reg(#NAME)
 
+class MyFactory {
+public:
+	static MyFactory* Instance() {
+		static MyFactory factory;
+		return &factory;
+	}
+
+	std::shared_ptr<NoisePlethoraPlugin> Create(std::string name) {
+		NoisePlethoraPlugin* instance = nullptr;
+
+		// find name in the registry and call factory method.
+		auto it = factoryFunctionRegistry.find(name);
+		if (it != factoryFunctionRegistry.end()) {
+			instance = it->second();
+		}
+
+		// wrap instance in a shared ptr and return
+		if (instance != nullptr) {
+			return std::shared_ptr<NoisePlethoraPlugin>(instance);
+		}
+		else {
+			return nullptr;
+		}
+	}
+
+	void RegisterFactoryFunction(std::string name, std::function<NoisePlethoraPlugin*(void)> classFactoryFunction) {
+		// register the class factory function
+		factoryFunctionRegistry[name] = classFactoryFunction;
+	}
+
+	std::map<std::string, std::function<NoisePlethoraPlugin*(void)>> factoryFunctionRegistry;
+};
+
+template<class T> class Registrar {
+public:
+	Registrar(std::string className) {
+		// register the class factory function
+		MyFactory::Instance()->RegisterFactoryFunction(className,
+		    [](void) -> NoisePlethoraPlugin * { return new T();});
+	}
+};
+
+#define REGISTER_PLUGIN(NAME) static Registrar<NAME> NAME ##_reg(#NAME)
