+++ src/noise-plethora/plugins/P_radioOhNo.hpp
@@ -7,16 +7,16 @@ class radioOhNo : public NoisePlethoraPlugin {
 public:
 
 	radioOhNo()
-		: patchCord1(dc1, 0, waveformMod1, 1),
-		  patchCord2(dc1, 0, waveformMod2, 1),
-		  patchCord3(dc1, 0, waveformMod3, 1),
-		  patchCord4(dc1, 0, waveformMod4, 1),
-		  patchCord5(waveformMod2, 0, waveformMod1, 0),
-		  patchCord6(waveformMod1, 0, waveformMod2, 0),
-		  patchCord7(waveformMod1, 0, mixer1, 0),
-		  patchCord8(waveformMod3, 0, waveformMod4, 0),
-		  patchCord9(waveformMod3, 0, mixer1, 1),
-		  patchCord10(waveformMod4, 0, waveformMod3, 0)
+		//: patchCord1(dc1, 0, waveformMod1, 1),
+		//  patchCord2(dc1, 0, waveformMod2, 1),
+		//  patchCord3(dc1, 0, waveformMod3, 1),
+		//  patchCord4(dc1, 0, waveformMod4, 1),
+		//  patchCord5(waveformMod2, 0, waveformMod1, 0),
+		//  patchCord6(waveformMod1, 0, waveformMod2, 0),
+		//  patchCord7(waveformMod1, 0, mixer1, 0),
+		//  patchCord8(waveformMod3, 0, waveformMod4, 0),
+		//  patchCord9(waveformMod3, 0, mixer1, 1),
+		//  patchCord10(waveformMod4, 0, waveformMod3, 0)
 	{ }
 
 	~radioOhNo() override {}
@@ -84,17 +84,16 @@ private:
 	AudioSynthWaveformModulated waveformMod3; //xy=468.75,729.75
 	AudioSynthWaveformModulated waveformMod4; //xy=470.75,848.75
 	AudioMixer4              mixer1;         //xy=676.750057220459,680.7500286102295
-	AudioConnection          patchCord1;
-	AudioConnection          patchCord2;
-	AudioConnection          patchCord3;
-	AudioConnection          patchCord4;
-	AudioConnection          patchCord5;
-	AudioConnection          patchCord6;
-	AudioConnection          patchCord7;
-	AudioConnection          patchCord8;
-	AudioConnection          patchCord9;
-	AudioConnection          patchCord10;
-
+	// AudioConnection          patchCord1;
+	// AudioConnection          patchCord2;
+	// AudioConnection          patchCord3;
+	// AudioConnection          patchCord4;
+	// AudioConnection          patchCord5;
+	// AudioConnection          patchCord6;
+	// AudioConnection          patchCord7;
+	// AudioConnection          patchCord8;
+	// AudioConnection          patchCord9;
+	// AudioConnection          patchCord10;
 };
 
 REGISTER_PLUGIN(radioOhNo);
