+++ src/noise-plethora/teensy/effect_multiply.h
@@ -32,5 +32,4 @@ class AudioEffectMultiply : public AudioStream {
 public:
 	AudioEffectMultiply() : AudioStream(2) { }
 	virtual void update(const audio_block_t* blocka, const audio_block_t* blockb, audio_block_t* blockout);
-private:
 };
